Let a stack be as large as it used to be
Upstream took any whole number from 1 upwards for maxStackSize;
this fork capped it at 100000, which was its own invention with
nothing behind it. The cap is gone - 1 to Integer.MAX_VALUE, for
the global value, the per-type ceilings and a region's override
alike. 0 stays refused, as upstream refused it; with 2147483647
available there is nothing left for it to mean.

The one sum that could overflow past a ceiling that large - "do
these two stacks together fit?" - is done in long now, so a count
can never wrap round to a negative and read as room to spare.

+++ common/src/main/java/com/frikinjay/mobstacker/MobStacker.java
@@ -361,7 +361,9 @@ public final class MobStacker {
             return false;
         }
 
-        if ((getStackSize(source) + getStackSize(target)) > getMaxMobStackSize(target)) {
+        // In long: a ceiling may be as large as an int goes, and two stacks near it would overflow
+        // to a negative and read as "plenty of room".
+        if (((long) getStackSize(source) + getStackSize(target)) > getMaxMobStackSize(target)) {
             return false;
         }
 
@@ -762,7 +764,8 @@ public final class MobStacker {
     }
 
     public static void mergeEntities(Mob target, Mob source) {
-        int newStackSize = Math.min(getStackSize(target) + getStackSize(source), getMaxMobStackSize(target));
+        int newStackSize = (int) Math.min((long) getStackSize(target) + getStackSize(source),
+                getMaxMobStackSize(target));
 
         // When two babies merge, keep the youngest (most negative) age so no member ever grows up
         // early — this replaces a strict age-band gate and lets baby-stacks freely consolidate.
